+++ src/test/java/it/miriade/runtime/core/RuntimeObjectFactoryTest.java
@@ -5,24 +5,24 @@ import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
-import it.miriade.runtime.core.RuntimeObjectFactory;
+import it.miriade.runtime.BaseTest;
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class RuntimeObjectFactoryTest {
+public class RuntimeObjectFactoryTest extends BaseTest {
 
 	private RuntimeObjectFactory<Object> factory;
 	private StringBuffer sourceCode = new StringBuffer();
 
 	@Test
 	public void t_00_build_demo() {
-		sourceCode.append("package test;");
-		sourceCode.append("public class Demo {");
-		sourceCode.append("	private String s;");
-		sourceCode.append("	private int i;");
-		sourceCode.append("	public String toString() {");
-		sourceCode.append("		return s + i;");
-		sourceCode.append("	}");
-		sourceCode.append("}");
+		sourceCode.append("package test;").append('\n');
+		sourceCode.append("public class Demo {").append('\n');
+		sourceCode.append("	private String s;").append('\n');
+		sourceCode.append("	private int i;").append('\n');
+		sourceCode.append("	public String toString() {").append('\n');
+		sourceCode.append("		return s + i;").append('\n');
+		sourceCode.append("	}").append('\n');
+		sourceCode.append("}").append('\n');
 
 		factory = new RuntimeObjectFactory<Object>("test.Demo", sourceCode.toString());
 		Assert.assertNotNull(factory);
@@ -33,18 +33,18 @@ public class RuntimeObjectFactoryTest {
 
 	@Test
 	public void t_01_build_demo_with_no_args() {
-		sourceCode.append("package test;");
-		sourceCode.append("public class Demo1 {");
-		sourceCode.append("	private String s;");
-		sourceCode.append("	private int i;");
-		sourceCode.append("	public Demo1() {");
-		sourceCode.append("		this.s = \"unknow\";");
-		sourceCode.append("		this.i = 0;");
-		sourceCode.append("	}");
-		sourceCode.append("	public String toString() {");
-		sourceCode.append("		return s + i;");
-		sourceCode.append("	}");
-		sourceCode.append("}");
+		sourceCode.append("package test;").append('\n');
+		sourceCode.append("public class Demo1 {").append('\n');
+		sourceCode.append("	private String s;").append('\n');
+		sourceCode.append("	private int i;").append('\n');
+		sourceCode.append("	public Demo1() {").append('\n');
+		sourceCode.append("		this.s = \"unknow\";").append('\n');
+		sourceCode.append("		this.i = 0;").append('\n');
+		sourceCode.append("	}").append('\n');
+		sourceCode.append("	public String toString() {").append('\n');
+		sourceCode.append("		return s + i;").append('\n');
+		sourceCode.append("	}").append('\n');
+		sourceCode.append("}").append('\n');
 
 		factory = new RuntimeObjectFactory<Object>("test.Demo1", sourceCode.toString());
 		Assert.assertNotNull(factory);
@@ -52,22 +52,21 @@ public class RuntimeObjectFactoryTest {
 		Assert.assertEquals("test.Demo1", t0.getClass().getName());
 		Assert.assertEquals("unknow0", t0.toString());
 	}
-	
 
 	@Test
 	public void t_02_build_demo_with_args() {
-		sourceCode.append("package test;");
-		sourceCode.append("public class Demo2 {");
-		sourceCode.append("	private String s;");
-		sourceCode.append("	private int i;");
-		sourceCode.append("	public Demo2(String s, int i) {");
-		sourceCode.append("		this.s = s;");
-		sourceCode.append("		this.i = i;");
-		sourceCode.append("	}");
-		sourceCode.append("	public String toString() {");
-		sourceCode.append("		return s + i;");
-		sourceCode.append("	}");
-		sourceCode.append("}");
+		sourceCode.append("package test;").append('\n');
+		sourceCode.append("public class Demo2 {").append('\n');
+		sourceCode.append("	private String s;").append('\n');
+		sourceCode.append("	private int i;").append('\n');
+		sourceCode.append("	public Demo2(String s, int i) {").append('\n');
+		sourceCode.append("		this.s = s;").append('\n');
+		sourceCode.append("		this.i = i;").append('\n');
+		sourceCode.append("	}").append('\n');
+		sourceCode.append("	public String toString() {").append('\n');
+		sourceCode.append("		return s + i;").append('\n');
+		sourceCode.append("	}").append('\n');
+		sourceCode.append("}").append('\n');
 
 		factory = new RuntimeObjectFactory<Object>("test.Demo2", sourceCode.toString());
 		Assert.assertNotNull(factory);