+++ src/test/java/it/miriade/runtime/core/runnables/MyRunnable.java
@@ -3,7 +3,7 @@ package it.miriade.runtime.core.runnables;
 import java.util.Arrays;
 import java.util.List;
 
-public interface MyRunnableInterface extends Runnable {
+public interface MyRunnable extends Runnable {
 
 	String first = "first";
 	String second = "second";